pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Вутятко <viktorzhirnov07@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.*"
fastapi = {extras = ["all"], version = "^0.86.0"}
pytest = "^7.2.0"
aioredis = "^2.0.1"
sqlmodel = "^0.0.8"
asyncpg = "^0.27.0"
pytest-asyncio = "0.20.1"
redis-om = "^0.1.0"
wemake-python-styleguide = "^0.17.0"
flake8-functions-names = "^0.3.0"
pep8-naming = "^0.13.2"
flake8-pep3101 = "^2.0.0"
flake8-use-fstring = "^1.4"
flake8-functions = "^0.0.7"
flake8-annotations-coverage = "^0.0.6"
flake8-annotations = "^2.9.1"
flake8-fastapi = "^0.7.0"
flake8-commas = "^2.1.0"
flake8-variables-names = "^0.0.5"
sqlalchemy2-stubs = "^0.0.2a29"
gunicorn = "^20.1.0"
pre-commit = "^2.20.0"
autoflake = "^1.7.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
